+++ otx/core/data/caching/__init__.py
@@ -5,5 +5,6 @@
 
 from .mem_cache_handler import MemCacheHandlerError, MemCacheHandlerSingleton
 from .mem_cache_hook import MemCacheHook
+from .storage_cache import init_arrow_cache
 
-__all__ = ["MemCacheHandlerSingleton", "MemCacheHook", "MemCacheHandlerError"]
+__all__ = ["MemCacheHandlerSingleton", "MemCacheHook", "MemCacheHandlerError", "init_arrow_cache"]
